CrowdWisdom trading/requirements.txt
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
pyyaml>=6.0
apify-client>=1.7.0
py-clob-client>=0.18.0
cryptography>=42.0.0
httpx>=0.27.0
ta>=0.11.0
rich>=13.7.0
colorlog>=6.8.0
scikit-learn>=1.4.0
streamlit>=1.35.0
plotly>=5.22.0
schedule>=1.2.2
